requirements.txt
tk
pillow
mysql-connector-python